Updated Admin Page

+++ controller/areaOfInterestTable.php
@@ -4,8 +4,8 @@ include("./model/areaOfInterestTable.php");
 
 $url = $_GET['url'];
 $url = explode('/',$url);
-if (isset($url[1])) {
-    $areaOfInterest->delete($url[1]);
+if (isset($url[1]) and !isset($url[2])) {
+  $areaOfInterest->delete($url[1]);
 }
 if ($url[1] == "add" and isset($_POST['submit'])) {
     $bg = $_POST['bg'];
@@ -19,6 +19,10 @@ if (isset($_POST['submit']) and isset($url[2])) {
   $areaOfInterest->update($url[2],$_POST['bg']);
   header("Location:../areaOfInterestTable");
 }
-$list = $areaOfInterest->selectAreaOfInterest();
-include("./view/areaOfInterestTable.php");
-?>
+$list = $areaOfInterest->selectBloodGroup();
+session_start();
+if(isset($_SESSION['name'])) {
+    include("./view/areaOfInterestTable.php");
+} else {
+  header("Location:login");
+}?>
