+++ controller/detailsOfGraduationTable.php
@@ -4,8 +4,8 @@ include("./model/detailsOfGraduationTable.php");
 
 $url = $_GET['url'];
 $url = explode('/',$url);
-if (isset($url[1])) {
-    $detailsOfGraduation->delete($url[1]);
+if (isset($url[1]) and !isset($url[2])) {
+  $detailsOfGraduation->delete($url[1]);
 }
 if ($url[1] == "add" and isset($_POST['submit'])) {
     $bg = $_POST['bg'];
@@ -19,6 +19,11 @@ if (isset($_POST['submit']) and isset($url[2])) {
   $detailsOfGraduation->update($url[2],$_POST['bg']);
   header("Location:../detailsOfGraduationTable");
 }
-$list = $detailsOfGraduation->selectDetailsOfGraduation();
-include("./view/detailsOfGraduationTable.php");
+$list = $detailsOfGraduation->selectBloodGroup();
+session_start();
+if(isset($_SESSION['name'])) {
+    include("./view/detailsOfGraduationTable.php");
+} else {
+  header("Location:login");
+}
 ?>
